move fake native networking to a separate file
