+++ src/www/app/cordova_main.ts
@@ -27,8 +27,9 @@ import * as sentry from '@sentry/browser';
 import {AbstractClipboard} from './clipboard';
 import {EnvironmentVariables} from './environment';
 import {SentryErrorReporter} from './error_reporter';
+import {FakeNativeNetworking} from './fake_net';
 import {main} from './main';
-import {FakeNativeNetworking, NativeNetworking} from './net';
+import {NativeNetworking} from './net';
 import {OutlinePlatform} from './platform';
 import {AbstractUpdater} from './updater';
 import * as interceptors from './url_interceptor';
